Performance & Serialization Enhancements (#20)
* feat: Add _ListAnnotation for inline JSON list serialization and enhance SchemaUtils for list handling

* fix: Improve asset handling in SchemaUtils by refining resource and array type checks

* feat: Enhance pickle serialization support in Packable class with a helper function

* feat: Add PackableWorker and worker entry point for offloading Packable reconstruction

* feat: Refactor extract_value method to use extract_basemodel for improved type annotation preservation

* feat: Optimize array encoding and decoding in ArrayUtils; enhance Packable serialization with orjson and caching

* Refactor meshly to use uint8 for index sizes and cell types

- Changed data types for index sizes and cell types from uint32 to uint8 in various classes and functions to optimize memory usage.
- Updated the ElementUtils and ElementData classes to handle the new data types and added a method to create ElementData from polygons.
- Modified the Mesh class to accommodate the new data types and adjusted related methods for consistency.
- Updated tests to reflect changes in data types, ensuring compatibility with the new structure.
- Adjusted TypeScript interfaces to align with the changes in the Python implementation, including updates to JSON schema definitions.

* feat: Integrate xxhash for improved checksum computation and add deterministic gzip compression

* Refactor mesh creation to use factory methods

- Updated all test files to replace direct instantiation of Mesh and its subclasses with the new factory method `create()`.
- Ensured that the conversion methods in tests now utilize the updated mesh creation approach.
- Adjusted tests for handling JAX arrays to align with the new instantiation method.
- Modified tests for dictionary arrays and dynamic models to reflect the changes in mesh creation.
- Ensured that all instances of `Mesh` and `CustomMesh` are created using the `create()` method for consistency and to leverage any internal optimizations.

* feat: Enhance Packable and SerializationUtils with checksum support and optimized extraction

* feat: Enhance SchemaUtils with nested model reconstruction and dynamic model instantiation

* feat: Update markers parameter to accept Nx3 numpy arrays and enhance handling in mesh creation

* refactor: Rename constants and update PackableStore to use a unified root directory structure

* feat: Implement PackableCache for efficient in-memory and disk caching of Packable objects

* refactor: Remove EXTRACTED_DIR constant and update PackableStore to use dynamic extracted directory

* feat: Add discriminator matching support in SchemaUtils for enhanced model resolution

* feat: Enhance PackableStore with root and extracted directory structure; add PackableCache for improved caching; update tests and documentation

* refactor: Replace xxhash with SHA256 for checksum computation and remove xxhash dependency

* feat: Add InlineArray type for inline JSON serialization; update related documentation and utilities

* refactor: Update extract_checksums method to be static and accept data dict as parameter

* chore: Bump version to 3.2.0-alpha in pyproject.toml and package.json

+++ python/tests/test_array_utils.py
@@ -64,7 +64,7 @@ class TestArrayUtils:
         test_arrays = [
             np.array([1, 2, 3, 4, 5], dtype=np.int32),
             np.array([1.1, 2.2, 3.3], dtype=np.float64),
-            np.array([[1, 2], [3, 4]], dtype=np.uint16),
+            np.array([[1, 2], [3, 4]], dtype=np.uint32),
             np.random.random((5, 5)).astype(np.float32),
         ]
